Make facet lists/nodes recursive
